src/test/java/pl/edu/wszib/kubalski/interpreter/expression/expressions/PiConstantExpressionTest.java
package pl.edu.wszib.kubalski.interpreter.expression.expressions;

import org.junit.jupiter.api.Test;
import pl.edu.wszib.kubalski.interpreter.Context;

import static org.junit.jupiter.api.Assertions.assertEquals;

class PiConstantExpressionTest {

    /**
     * Test class for PiExpression, which represents a constant expression
     * returning the mathematical constant π (pi).
     * <p>
     * Method being tested: interpret(Context context).
     */
    @Test
    void shouldReturnMathPIWhenInterpretIsCalled() {
        // Arrange
        PiConstantExpression piExpression = new PiConstantExpression();
        Context context = Context.builder().build();

        // Act
        Double result = piExpression.interpret(context);

        // Assert
        assertEquals(Math.PI, result, "The interpret method should return the value of PI");
    }
}
